5286 client - Fix false duplicate node name warning for nested tasks
getDuplicateNodeNames received the flattened task list (Workflow.getTasks(true)),
where every nested task already appears as a top-level entry while its dispatcher
still carries it in parameters. Recursing into nested groups counted each nested
task twice and falsely flagged it as a duplicate for any workflow with a nested
task. Count only trigger and top-level task names; the flattened list already
contains every node once.
